requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distlib==0.3.9
filelock==3.18.0
flake8==7.2.0
flake8-bugbear==24.12.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.9
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
piexif==1.1.3
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
pycodestyle==2.13.0
pydantic==2.11.1
pydantic_core==2.33.0
pyflakes==3.3.2
pytest==8.3.5
python-telegram-bot==22.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
virtualenv==20.29.3
